removed validator keeper functions
